+++ src/main/java/devcourse/springbootbasic/controller/VoucherController.java
@@ -1,10 +1,14 @@
 package devcourse.springbootbasic.controller;
 
+import devcourse.springbootbasic.dto.customer.CustomerFindResponse;
 import devcourse.springbootbasic.dto.voucher.VoucherCreateRequest;
 import devcourse.springbootbasic.dto.voucher.VoucherFindResponse;
 import devcourse.springbootbasic.dto.voucher.VoucherResponse;
 import devcourse.springbootbasic.dto.voucher.VoucherUpdateDiscountValueRequest;
+import devcourse.springbootbasic.dto.wallet.VoucherAssignRequest;
+import devcourse.springbootbasic.dto.wallet.VoucherAssignResponse;
 import devcourse.springbootbasic.service.VoucherService;
+import devcourse.springbootbasic.service.WalletService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 
@@ -16,6 +20,7 @@ import java.util.UUID;
 public class VoucherController {
 
     private final VoucherService voucherService;
+    private final WalletService walletService;
 
     public List<VoucherFindResponse> findAllVouchers() {
         return this.voucherService.findAllVouchers();
@@ -31,5 +36,17 @@ public class VoucherController {
 
     public VoucherResponse deleteVoucher(UUID voucherId) {
         return new VoucherResponse(this.voucherService.deleteVoucher(voucherId));
+    }
+
+    public VoucherAssignResponse assignVoucherToCustomer(VoucherAssignRequest voucherAssignRequest) {
+        return this.walletService.assignVoucherToCustomer(voucherAssignRequest);
+    }
+
+    public VoucherAssignResponse unassignVoucherFromCustomer(UUID voucherId) {
+        return this.walletService.unassignVoucherFromCustomer(voucherId);
+    }
+
+    public CustomerFindResponse findCustomerByVoucherId(UUID voucherId) {
+        return this.walletService.findCustomerByVoucherId(voucherId);
     }
 }
